requirements.txt
ollama>=0.1.6
matplotlib>=3.8.3
numpy>=1.26.4
pandas>=2.2.1
psutil>=5.9.8
GPUtil>=1.4.0
py-cpuinfo>=9.0.0
setuptools>=69.0.0
